GPU: Rename BlendFunc's enum value None to NoBlend
It looks like None is defined in X.h as 0L so breaking
compilation on platforms which use X. Instead of playing
with the preprocessor just rename the enum value.
As seen on:
http://forums.desmume.org/viewtopic.php?pid=20591

Thanks Alberto for forwarding the issue.

Remember: everytime a bug is reported in a forum a kitten dies!


git-svn-id: http://svn.code.sf.net/p/desmume/code/trunk@4535 5577840e-afb3-4e59-9ef3-2b7f898b1952